+++ api-derive/bagsList/listNodes.ts
@@ -2,14 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.6';
-import type { Option } from 'https://deno.land/x/polkadot@0.2.6/types/mod.ts';
-import type { AccountId32 } from 'https://deno.land/x/polkadot@0.2.6/types/interfaces/index.ts';
-import type { PalletBagsListListBag, PalletBagsListListNode } from 'https://deno.land/x/polkadot@0.2.6/types/lookup.ts';
+import type { Option } from 'https://deno.land/x/polkadot/types/mod.ts';
+import type { AccountId32 } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { PalletBagsListListBag, PalletBagsListListNode } from 'https://deno.land/x/polkadot/types/lookup.ts';
 import type { DeriveApi } from '../types.ts';
 
 import { BehaviorSubject, map, of, switchMap, tap, toArray } from 'https://esm.sh/rxjs@7.5.6';
 
-import { nextTick } from 'https://deno.land/x/polkadot@0.2.6/util/mod.ts';
+import { nextTick } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { memo } from '../util/index.ts';