+++ types-codec/native/Json.ts
@@ -1,10 +1,10 @@
 // Copyright 2017-2022 @polkadot/types-codec authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { HexString } from 'https://deno.land/x/polkadot@0.2.6/util/types.ts';
+import type { HexString } from 'https://deno.land/x/polkadot/util/types.ts';
 import type { AnyJson, Codec, Inspect, IU8a, Registry } from '../types/index.ts';
 
-import { isFunction, objectProperties, stringify } from 'https://deno.land/x/polkadot@0.2.6/util/mod.ts';
+import { isFunction, objectProperties, stringify } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { compareMap } from '../utils/index.ts';