+++ types-known/spec/polkadot.ts
@@ -3,9 +3,9 @@
 
 /* eslint-disable sort-keys */
 
-import type { OverrideVersionedType } from 'https://deno.land/x/polkadot@0.2.6/types/types/index.ts';
+import type { OverrideVersionedType } from 'https://deno.land/x/polkadot/types/types/index.ts';
 
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.6/util/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 const sharedTypes = {
   CompactAssignments: 'CompactAssignmentsWith16',
@@ -23,7 +23,8 @@ const sharedTypes = {
       CancelProxy: 6,
       Auction: 7
     }
-  }
+  },
+  Weight: 'u64'
 };
 
 const addrAccountIdTypes = {