+++ types-known/spec/rococo.ts
@@ -3,10 +3,10 @@
 
 /* eslint-disable sort-keys */
 
-import type { OverrideVersionedType } from 'https://deno.land/x/polkadot@0.2.6/types/types/index.ts';
+import type { OverrideVersionedType } from 'https://deno.land/x/polkadot/types/types/index.ts';
 
-import { mapXcmTypes } from 'https://deno.land/x/polkadot@0.2.6/types-create/mod.ts';
-import { objectSpread } from 'https://deno.land/x/polkadot@0.2.6/util/mod.ts';
+import { mapXcmTypes } from 'https://deno.land/x/polkadot/types-create/mod.ts';
+import { objectSpread } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 // structs need to be in order
 /* eslint-disable sort-keys */
@@ -14,7 +14,8 @@ import { objectSpread } from 'https://deno.land/x/polkadot@0.2.6/util/mod.ts';
 const sharedTypes = {
   DispatchErrorModule: 'DispatchErrorModuleU8',
   FullIdentification: '()', // No staking, only session (as per config)
-  Keys: 'SessionKeys7B'
+  Keys: 'SessionKeys7B',
+  Weight: 'u64'
 };
 
 const versioned: OverrideVersionedType[] = [